undo the test moves; user management testing still needs them duplicated
